requirements.txt
# ViralEvo Python dependencies
# Install: pip install -r requirements.txt

requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0

# Standard library (no install needed):
# sqlite3, json, pathlib, datetime, uuid, argparse, re, subprocess
